split impl and def for file reader
