+++ src/main.cpp
@@ -1,5 +1,7 @@
 #include "file_reader.h"
 
+#include <iostream>
+
 int main(int argc, const char* argv[])
 {
 	if (argc != 2)
